+++ golang/lib/networks/network.go
@@ -17,5 +17,5 @@
 
 package networks
 
-// Docs available at https://docs.kurtosistech.com/kurtosis-libs/lib-documentation
+// Docs available at https://docs.kurtosistech.com/kurtosis-client/lib-documentation
 type Network interface{}