+++ docs/changelog.md
@@ -1,5 +1,13 @@
 # TBD
 
+# 0.14.0
+### Breaking Changes
+* Split method `NetworkContext.WaitForEndpointAvailability` to `NetworkContext.WaitForHttpGetEndpointAvailability` and `NetworkContext.WaitForHttpPostEndpointAvailability`
+  * Users should replace their `NetworkContext.WaitForEndpointAvailability` calls with `NetworkContext.WaitForHttpGetEndpointAvailability` or `NetworkContext.WaitForHttpPostEndpointAvailability` depending on the endpoint used to check availability
+
+### Fixes
+* Fixed broken documentation links in code
+
 # 0.13.7
 ### Changes
 * Abolished the use of ResultAsync, okAsync, and errAsync in favour of just regular Result